base-passwd: copy base-passwd's master files into staging's sysconfdir just once
base-passwd should not overwrite the /etc/passwd and /etc/groups on the staging
area because other useradd recipes may have already include items. This follows
the same logic as populate_packages_prepend function defined in the same recipe.

[YOCTO #7724]

+++ meta/recipes-core/base-passwd/base-passwd_3.5.29.bb
@@ -57,9 +57,13 @@ base_passwd_sstate_postinst() {
 		# them to make sure they are atomically install.
 		install -d -m 755 ${STAGING_DIR_TARGET}${sysconfdir}
 		for i in passwd group; do
-			install -p -m 644 ${STAGING_DIR_TARGET}${datadir}/base-passwd/$i.master \
-				${STAGING_DIR_TARGET}${sysconfdir}/
-			mv ${STAGING_DIR_TARGET}${sysconfdir}/$i.master ${STAGING_DIR_TARGET}${sysconfdir}/$i
+			# Copy the ${sysconfdir} files only the first time, so we do not overwrite
+			# users/group created by useradd recipes
+			if [ ! -e ${STAGING_DIR_TARGET}${sysconfdir}/$i ]; then
+				install -p -m 644 ${STAGING_DIR_TARGET}${datadir}/base-passwd/$i.master \
+					${STAGING_DIR_TARGET}${sysconfdir}/
+				mv ${STAGING_DIR_TARGET}${sysconfdir}/$i.master ${STAGING_DIR_TARGET}${sysconfdir}/$i
+			fi
 		done
 	fi
 }
